+++ ColorPicker/ColorPicker.js
@@ -1,4 +1,21 @@
 /* eslint-disable react-hooks/rules-of-hooks */
+/**
+ * Sandstone component to allow the user to choose a color.
+ *
+ * @example
+ * const [color, setColor]
+ * <ColorPicker
+ *	 color={'#FF00FF'}
+ *	 colorHandler={setColor}
+ *	 presetColors={['#FF0000', '#00FF00', '#0000FF']}
+ *	 text={'Color Picker'}
+ * />
+ *
+ * @module sandstone/ColorPicker
+ * @exports ColorPicker
+ * @exports ColorPickerBase
+ * @exports ColorPickerDecorator
+ */
 
 import kind from '@enact/core/kind';
 import Spottable from '@enact/spotlight/Spottable';
@@ -16,104 +33,9 @@ import Popup from '../Popup';
 import Skinnable from '../Skinnable';
 import Slider from '../Slider';
 
+import {hexToHSL, HSLToHex} from './utils';
+
 import componentCss from './ColorPicker.module.less';
-
-const checkHex = (hex) => {
-	if (/^#[0-9A-F]{6}$/i.test(hex)) return hex;
-	return '#000000';
-};
-const hexToHSL = (hexColor) => {
-	const hex = checkHex(hexColor);
-	// Convert hex to RGB first
-	let r = 0, g = 0, b = 0;
-	if (hex.length === 4) {
-		r = parseInt(hex[1] + hex[1], 16);
-		g = parseInt(hex[2] + hex[2], 16);
-		b = parseInt(hex[3] + hex[3], 16);
-	} else if (hex.length === 7) {
-		r = parseInt(hex.slice(1, 3), 16);
-		g = parseInt(hex.slice(3, 5), 16);
-		b = parseInt(hex.slice(5), 16);
-	}
-
-	// Then convert RGB to HSL
-	r /= 255;
-	g /= 255;
-	b /= 255;
-	let cmin = Math.min(r, g, b),
-		cmax = Math.max(r, g, b),
-		delta = cmax - cmin,
-		h, s, l;
-
-	if (delta === 0) {
-		h = 0;
-	} else if (cmax === r) {
-		h = ((g - b) / delta) % 6;
-	} else if (cmax === g) {
-		h = (b - r) / delta + 2;
-	} else {
-		h = (r - g) / delta + 4;
-	}
-
-	h = Math.round(h * 60);
-
-	if (h < 0) {
-		h += 360;
-	}
-
-	l = (cmax + cmin) / 2;
-	s = delta === 0 ? 0 : delta / (1 - Math.abs(2 * l - 1));
-
-	s = +(s * 100).toFixed(1);
-	l = +(l * 100).toFixed(1);
-
-	return {h: Math.round(h), s: Math.round(s), l: Math.round(l)};
-};
-
-const HSLToHex = (h, s, l) => {
-	s /= 100;
-	l /= 100;
-
-	let c = (1 - Math.abs(2 * l - 1)) * s,
-		x = c * (1 - Math.abs((h / 60) % 2 - 1)),
-		m = l - c / 2,
-		r = 0,
-		g = 0,
-		b = 0;
-
-	if (0 <= h && h < 60) {
-		r = c; g = x; b = 0;
-	} else if (60 <= h && h < 120) {
-		r = x; g = c; b = 0;
-	} else if (120 <= h && h < 180) {
-		r = 0; g = c; b = x;
-	} else if (180 <= h && h < 240) {
-		r = 0; g = x; b = c;
-	} else if (240 <= h && h < 300) {
-		r = x; g = 0; b = c;
-	} else if (300 <= h && h < 360) {
-		r = c; g = 0; b = x;
-	}
-	// Having obtained RGB, convert channels to hex
-	r = Math.round((r + m) * 255).toString(16);
-	g = Math.round((g + m) * 255).toString(16);
-	b = Math.round((b + m) * 255).toString(16);
-
-	// Prepend 0s, if necessary
-	if (r.length === 1) {
-		r = "0" + r;
-	}
-
-	if (g.length === 1) {
-		g = "0" + g;
-	}
-
-	if (b.length === 1) {
-		b = "0" + b;
-	}
-
-	return "#" + r + g + b;
-};
 
 const SpottableButton = Spottable(ButtonBase);
 
@@ -189,7 +111,7 @@ const PopupContent = ({color, colorHandler, css, presetColors}) => {
 						onChange={changeHue}
 						value={hue}
 					/>
-					<BodyText className={css.colorSliderText} css={css}>Saturation %{saturation}</BodyText>
+					<BodyText className={css.colorSliderText} css={css}>Saturation {saturation}%</BodyText>
 					<Slider
 						className={css.colorSlider}
 						max={100}
@@ -199,7 +121,7 @@ const PopupContent = ({color, colorHandler, css, presetColors}) => {
 						onChange={changeSaturation}
 						value={saturation}
 					/>
-					<BodyText className={css.colorSliderText} css={css}>Lightness %{lightness}</BodyText>
+					<BodyText className={css.colorSliderText} css={css}>Lightness {lightness}%</BodyText>
 					<Slider
 						className={css.colorSlider}
 						max={100}
@@ -257,7 +179,7 @@ PopupContent.propTypes = {
 };
 
 /**
- * A color picker component.
+ * The color picker base component which sets-up the component's structure.
  *
  * This component is most often not used directly but may be composed within another component as it
  * is within {@link sandstone/ColorPicker|ColorPicker}.
@@ -372,6 +294,8 @@ const ColorPickerBase = kind({
 	},
 
 	render: ({color, colorHandler, css, disabled, handleClosePopup, handleOpenPopup, popupOpen, presetColors, text, ...rest}) => {
+		delete rest.onTogglePopup;
+
 		const CloseIcon = useCallback((props) => <Icon {...props} css={css} />, [css]);
 		const slotAfter = <SpottableButton
 			className={css.coloredButton}
@@ -408,7 +332,6 @@ const ColorPickerBase = kind({
 		);
 	}
 });
-
 
 /**
  * Applies Sandstone specific behaviors to {@link sandstone/ColorPicker.ColorPickerBase|ColorPicker} components.