+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `sandstone/ColorPicker` component
+
 ## [2.7.3] - 2023-07-14
 
 ### Fixed